refactor(web): remove fleet sidebar and legacy session surface
